feat: Reorganized widget components
