+++ apps/widget/craco.config.js
@@ -1,5 +1,10 @@
+const path = require('path');
 module.exports = {
-  babel: {
-    plugins: [],
+  webpack: {
+    alias: {
+      "@store": path.resolve(__dirname, "./src/store"),
+      "@client": path.resolve(__dirname, "./src/client"),
+      "@config": path.resolve(__dirname, "./src/config"),
+    },
   },
 };
