+++ apps/widget/src/components/widget/Container/index.ts
@@ -1,0 +1,1 @@
+export * from './Container';
