+++ apps/widget/src/components/widget/Provider/index.ts
@@ -1,0 +1,1 @@
+export * from './Provider';
