--- apps/widget/src/components/widget/WidgetContainer.tsx
@@ -1,8 +1,0 @@
-import { Widget } from './Widget';
-import { useParams } from 'react-router-dom';
-
-export function NotificationCenterWidgetContainer() {
-  const { projectId = '' } = useParams<{ projectId: string }>();
-
-  return <Widget projectId={projectId} />;
-}
