+++ apps/widget/src/components/widget/index.ts
@@ -1,2 +1,1 @@
 export * from './Widget';
-export * from './WidgetContainer';
